Cleaned up warnings.

+++ platform/xcb/screen.cpp
@@ -28,7 +28,7 @@ screen::~screen( void )
 draw::size screen::bounds( void )
 {
 	precondition( _screen, "null screen" );
-	return { _screen->width_in_pixels, _screen->height_in_pixels };
+	return { double(_screen->width_in_pixels), double(_screen->height_in_pixels) };
 }
 
 ////////////////////////////////////////
